Make viewport height adjusted
