+++ examples/examples3D/suzanne_example.cpp
@@ -1,6 +1,6 @@
 #include "Vivid.h"
 InputHandler* InputHandler::s_Instance;
-Vivid::Camera* Vivid::Camera::s_Instance;
+Camera* Camera::s_Instance;
 
 class ExampleInterface : public RenderingInterface
 {
@@ -9,7 +9,12 @@ private:
 	Vec3 lightPos = Vec3(0.0f, 0.0f, -100.0f);
 	Vec2* m_PrevMousePosition = new Vec2(0.0f, 0.0f);
 
-public:
+	glm::vec3 translationModel1 = glm::vec3(0, 50, -200);
+	Vivid::Mesh lightMesh;
+	Vivid::PointLight light;
+	Ref<Vivid::Shader> lightShader;
+
+	    public:
 	void Setup() override
 	{
 		// Can write custom opengl confs here
@@ -19,9 +24,9 @@ public:
 
 		// Light Info
 		Vivid::Quad3d cube(10.0f, lightColor);
-		Vivid::Mesh lightMesh(cube);
-		Vivid::PointLight light(lightPos, lightColor, &lightMesh);
-		Ref<Vivid::Shader> lightShader = MakeRef<Vivid::Shader>("./../assets/shaders/basic.vertexShader.glsl",
+		lightMesh = Vivid::Mesh(cube);
+		light = Vivid::PointLight(lightPos, lightColor, &lightMesh);
+		lightShader = MakeRef<Vivid::Shader>("./../assets/shaders/basic.vertexShader.glsl",
 		    "./../assets/shaders/basic.pixelShader.glsl");
 		light.Draw(lightShader);
 
@@ -33,7 +38,6 @@ public:
 		Vivid::Mesh mesh1("./../assets/obj/suzanne.obj");
 		mesh1.BindShader(shader);
 
-		glm::vec3 translationModel1(0, 50, -200);
 
 		mesh1.Update(glm::translate(glm::mat4(1.0f), translationModel1));
 		shader->SetUniform3f("lightColor", lightColor);
@@ -57,8 +61,6 @@ public:
 		Vivid::Mesh mesh1("./../assets/obj/suzanne.obj");
 		mesh1.BindShader(shader);
 
-		glm::vec3 translationModel1(0, 50, -200);
-
 		shader->SetUniform3f("lightColor", lightColor);
 		shader->SetUniform3f("lightPos", lightPos);
 		shader->SetUniform1f("intensity", 2.0f);
@@ -66,16 +68,20 @@ public:
 		shader->Bind();
 		mesh1.Update(glm::translate(glm::mat4(1.0f), translationModel1));
 		mesh1.Draw();
+
+		light.UpdateLightPosition(lightPos);
+		light.Draw(lightShader);
 	}
 
 	void ImGuiRender() override
 	{
 		ImGui::Begin("Debug");
-		//	ImGui::SliderFloat3("Translation Model 1", &translationModel1.x, -500.0f, 500.0f);
-		//	// ImGui::SliderFloat3("Translation Model 2", &translationModel2.x, -300.0f, 300.0f);
-		//	ImGui::SliderFloat3("Light Position", &lightPos.x, -500.0f, 500.0f);
+		ImGui::SliderFloat3("Translation Model 1", &translationModel1.x, -500.0f, 500.0f);
+		ImGui::SliderFloat3("Light Position", &lightPos.x, -500.0f, 500.0f);
+
 		ImGui::Text("Application average %.3f ms/frame (%.1f FPS)",
 		    1000.0f / ImGui::GetIO().Framerate, ImGui::GetIO().Framerate);
+
 
 		ImGui::End();
 	}
